commande fournisseur finished

+++ src/main/java/com/practice/gestion_st/Dao/ProduitRepository.java
@@ -2,6 +2,6 @@ package com.practice.gestion_st.Dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.practice.gestion_st.Entity.Produit;
 
-public interface ProduitRepository extends JpaRepository<Produit, Integer>{
+public interface ProduitRepository extends JpaRepository<Produit, Long>{
 		
 }
