+++ src/main/java/com/practice/gestion_st/entity/Produit.java
@@ -1,10 +1,6 @@
 package com.practice.gestion_st.Entity;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
+import javax.persistence.*;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -14,7 +10,7 @@ import lombok.NoArgsConstructor;
 @Data @NoArgsConstructor @AllArgsConstructor
 public class Produit {
 	@Id @GeneratedValue
-	private int id;
+	private long id;
 	private String designation;
 	private String image;
 	private Long qnt_stocke;
